Update index.js
Resolved a problem that occurred when using devices with multiple channels, like the Sonoff 4CH or the 2CH. There was an error saving the correct device id as key for the local cache. The result was that plugin try to create a new accessories with the classic error "Cannot add a bridged Accessory with the same UUID as another bridged Accessory".